requirements.txt
Django==1.11.11
django-countries==5.3
django-widget-tweaks==1.4.2
django-solo==1.1.3
reportlab==3.4.0
xhtml2pdf==0.2.2